requirements.txt
asgiref==3.5.2
Django==4.0.4
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
packaging==21.3
pyparsing==3.0.9
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0